Respect the html_use_smartypants config option
Previously, this was ignored because the config option is implemented
by using a different HTMLTranslator in the case that the option is
True (which is the default). This PR switches from subclassing the
non-SmartyPants HTMLTranslator and instead uses calls to
`app.add_node` to override the visit/depart functions.

This removes the need to use `app.set_translator`, making our
Sphinx requirement more flexible. I tested building the demo with
the current version of Sphinx and earlier, and it worked all the
way to Sphinx 1.1, which failed for something unrelated to this change.
`setup.py` was updated accordingly.